Bring AHP in line with Hacklab Conflict of Interest policies.
Previously, in Hacklab's Anti Harrassment Policy, the role of
deciding whether to investigate an allged harrassment and performing the
investigation rested with Hacklab's President. This is clearly
problematic in the scenario where the president is the victim or alleged
perpetrator. It is inconsistent with the spirit of Hacklab's Conflict
of Interest policies.

To corect for this, we split the President's role in AHP investigations
into three parts and allow for them to be handled by any Director or
Officer.

The first role, that of the Contacted Person, is an Officer or Director
the complainant feels comfortable contacting. They do not need to be
impartial, as their sole role is to bring the issue to the Board.

The second role, that of deciding whether to investigate the issue, is
placed with Hacklab's Board of Directors. (Persuant to the Conflict of
Interest Policies, Directors in a conflict of interest are absent.)

The third role, that of the investigator, is given to an impartial
individual selected by the Board.